Traveling salesman problem
We solved a variant of the travelling salesman problem (TSP) using the A* search algorithm with the Manhattan distance heuristic. We construct a shortest path graph by calculating the pair-wise distance for all checkpoints on the map, and then we find the shortest way with BFS and UCS for the robot to visit all the checkpoints exactly once and return to its starting checkpoint.